requirements.txt
CommonMark
gnupg
pytest
python-magic
